+++ esm/custom-element-registry.js
@@ -4,6 +4,8 @@ const {entries, setPrototypeOf} = Object;
 
 let reactive = false;
 
+export const getReactive = () => reactive;
+
 export const setReactive = value => {
   reactive = value;
 };
