+++ package.json
@@ -39,6 +39,10 @@
       "import": "./esm/index.js",
       "default": "./cjs/index.js"
     },
+    "./cached": {
+      "import": "./esm/cached.js",
+      "default": "./cjs/cached.js"
+    },
     "./package.json": "./package.json"
   },
   "dependencies": {
