+++ test/benchmark-linkedom.js
@@ -1,5 +1,5 @@
 const benchmark = require('./benchmark.js');
-const {DOMParser} = require('../cjs');
+const {DOMParser} = require('../cjs/cached');
 const dp = new DOMParser;
 
 benchmark('linkedom', html => dp.parseFromString(html, 'text/html'));
